--- crates/fbuild-config/src/ini_parser.rs
@@ -1,1603 +1,0 @@
-//! PlatformIO INI parser with environment inheritance and variable substitution.
-//!
-//! Features:
-//! - Sections: `[env:name]`, `[platformio]`, custom sections
-//! - `extends` directive for environment inheritance
-//! - Variable substitution: `${section.key}` or `${env:section.key}`
-//! - Multi-line values (continuation lines starting with whitespace)
-//! - Inline comments (` ; comment` or ` # comment`)
-//! - Base `[env]` section merged into all `[env:*]` sections
-
-use std::collections::HashMap;
-use std::collections::HashSet;
-use std::path::{Path, PathBuf};
-
-use regex::Regex;
-
-use crate::pio_env::PioEnvOverrides;
-
-/// Parsed platformio.ini configuration.
-pub struct PlatformIOConfig {
-    /// Raw sections: section_name -> key -> value
-    sections: HashMap<String, HashMap<String, String>>,
-    /// Resolved environment configs (with inheritance applied)
-    resolved_envs: HashMap<String, HashMap<String, String>>,
-    /// Path to the platformio.ini file
-    path: PathBuf,
-    /// Per-request `PLATFORMIO_*` env var overrides forwarded from the caller.
-    ///
-    /// Used by getters to honor env-driven overrides without reading
-    /// `std::env::var` directly. The daemon does not inherit caller env vars,
-    /// so all `PLATFORMIO_*` config must flow through this struct.
-    overrides: PioEnvOverrides,
-}
-
-impl PlatformIOConfig {
-    /// Parse a platformio.ini file with no env var overrides.
-    ///
-    /// Equivalent to `from_path_with_overrides(path, PioEnvOverrides::empty())`.
-    pub fn from_path(path: &Path) -> fbuild_core::Result<Self> {
-        Self::from_path_with_overrides(path, PioEnvOverrides::empty())
-    }
-
-    /// Parse a platformio.ini file and attach `PLATFORMIO_*` env var overrides.
-    ///
-    /// The overrides are consulted by getters before falling back to INI values,
-    /// allowing CLI callers to forward env vars to the daemon over HTTP without
-    /// the daemon process needing to inherit them.
-    pub fn from_path_with_overrides(
-        path: &Path,
-        overrides: PioEnvOverrides,
-    ) -> fbuild_core::Result<Self> {
-        let content = std::fs::read_to_string(path).map_err(|e| {
-            fbuild_core::FbuildError::ConfigError(format!(
-                "failed to read {}: {}",
-                path.display(),
-                e
-            ))
-        })?;
-
-        let sections = Self::parse_ini(&content)?;
-        let resolved_envs = Self::resolve_all_envs(&sections)?;
-
-        Ok(Self {
-            sections,
-            resolved_envs,
-            path: path.to_path_buf(),
-            overrides,
-        })
-    }
-
-    /// Borrow the env var overrides attached to this config.
-    pub fn overrides(&self) -> &PioEnvOverrides {
-        &self.overrides
-    }
-
-    /// List all environment names (from `[env:name]` sections).
-    pub fn get_environments(&self) -> Vec<&str> {
-        let mut envs: Vec<&str> = self.resolved_envs.keys().map(|s| s.as_str()).collect();
-        envs.sort();
-        envs
-    }
-
-    /// Check if an environment exists.
-    pub fn has_environment(&self, env_name: &str) -> bool {
-        self.resolved_envs.contains_key(env_name)
-    }
-
-    /// Get resolved config for an environment (with inheritance applied).
-    pub fn get_env_config(&self, env_name: &str) -> fbuild_core::Result<&HashMap<String, String>> {
-        self.resolved_envs.get(env_name).ok_or_else(|| {
-            fbuild_core::FbuildError::ConfigError(format!("environment '{}' not found", env_name))
-        })
-    }
-
-    /// Get the default environment name.
-    ///
-    /// Priority:
-    /// 1. forwarded `PLATFORMIO_DEFAULT_ENVS` override (first value)
-    /// 2. `[platformio]` section `default_envs` key (first value)
-    /// 3. First environment in file order
-    /// 4. None if no environments
-    pub fn get_default_environment(&self) -> Option<&str> {
-        // Check forwarded PLATFORMIO_DEFAULT_ENVS first.
-        if let Some(defaults) = self.overrides.get_default_envs() {
-            let first = defaults.split(',').next().unwrap_or("").trim();
-            if !first.is_empty() && self.resolved_envs.contains_key(first) {
-                return Some(
-                    self.resolved_envs
-                        .keys()
-                        .find(|k| k.as_str() == first)
-                        .map(|k| k.as_str())
-                        .unwrap(),
-                );
-            }
-        }
-
-        // Fall back to [platformio].default_envs.
-        if let Some(pio) = self.sections.get("platformio") {
-            if let Some(defaults) = pio.get("default_envs") {
-                let first = defaults.split(',').next().unwrap_or("").trim();
-                if !first.is_empty() && self.resolved_envs.contains_key(first) {
-                    return Some(
-                        self.resolved_envs
-                            .keys()
-                            .find(|k| k.as_str() == first)
-                            .map(|k| k.as_str())
-                            .unwrap(),
-                    );
-                }
-            }
-        }
-        // Fall back to first environment
-        let mut envs: Vec<&str> = self.resolved_envs.keys().map(|s| s.as_str()).collect();
-        envs.sort();
-        envs.into_iter().next()
-    }
-
-    /// Get build flags for an environment, parsed into a list.
-    ///
-    /// Handles:
-    /// - Space-separated flags on one line
-    /// - Multi-line flags (one per line)
-    /// - `-D FLAG` → `-DFLAG` normalization
-    pub fn get_build_flags(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        if let Some(flags) = self.overrides.get_build_flags() {
-            return Ok(parse_flags(flags));
-        }
-
-        let config = self.get_env_config(env_name)?;
-        match config.get("build_flags") {
-            Some(flags) => Ok(parse_flags(flags)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get build_src_flags for an environment (sketch-only flags).
-    pub fn get_build_src_flags(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        if let Some(flags) = self.overrides.get_build_src_flags() {
-            return Ok(parse_flags(flags));
-        }
-
-        let config = self.get_env_config(env_name)?;
-        match config.get("build_src_flags") {
-            Some(flags) => Ok(parse_flags(flags)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get build_unflags for an environment.
-    ///
-    /// Priority:
-    /// 1. `PLATFORMIO_BUILD_UNFLAGS` forwarded override
-    /// 2. `build_unflags`
-    pub fn get_build_unflags(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        if let Some(flags) = self.overrides.get_build_unflags() {
-            return Ok(parse_flags(flags));
-        }
-
-        let config = self.get_env_config(env_name)?;
-        match config.get("build_unflags") {
-            Some(flags) => Ok(parse_flags(flags)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get build_type for an environment.
-    ///
-    /// PlatformIO defaults this to `release`.
-    pub fn get_build_type(&self, env_name: &str) -> fbuild_core::Result<String> {
-        let config = self.get_env_config(env_name)?;
-        Ok(config
-            .get("build_type")
-            .map(|value| strip_inline_comment(value))
-            .filter(|value| !value.is_empty())
-            .unwrap_or_else(|| "release".to_string()))
-    }
-
-    /// Get debug_build_flags for an environment.
-    ///
-    /// PlatformIO defaults to `-Og -g2 -ggdb2` when the option is not set.
-    pub fn get_debug_build_flags(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        let config = self.get_env_config(env_name)?;
-        match config.get("debug_build_flags") {
-            Some(flags) => Ok(parse_flags(flags)),
-            None => Ok(vec![
-                "-Og".to_string(),
-                "-g2".to_string(),
-                "-ggdb2".to_string(),
-            ]),
-        }
-    }
-
-    /// Get source filter rules for an environment.
-    ///
-    /// Priority:
-    /// 1. `PLATFORMIO_BUILD_SRC_FILTER` forwarded override
-    /// 2. `build_src_filter`
-    /// 3. legacy `src_filter`
-    pub fn get_source_filter(&self, env_name: &str) -> fbuild_core::Result<Option<String>> {
-        if let Some(value) = self.overrides.get_build_src_filter() {
-            let cleaned = strip_inline_comment(value);
-            if !cleaned.is_empty() {
-                return Ok(Some(cleaned.to_string()));
-            }
-        }
-
-        let config = self.get_env_config(env_name)?;
-        for key in ["build_src_filter", "src_filter"] {
-            if let Some(value) = config.get(key) {
-                let cleaned = strip_inline_comment(value);
-                if !cleaned.is_empty() {
-                    return Ok(Some(cleaned.to_string()));
-                }
-            }
-        }
-
-        Ok(None)
-    }
-
-    /// Get library dependencies for an environment.
-    pub fn get_lib_deps(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        let config = self.get_env_config(env_name)?;
-        match config.get("lib_deps") {
-            Some(deps) => Ok(parse_lib_deps(deps)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get lib_ignore for an environment (libraries to skip).
-    pub fn get_lib_ignore(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        let config = self.get_env_config(env_name)?;
-        match config.get("lib_ignore") {
-            Some(deps) => Ok(parse_lib_deps(deps)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get `extra_scripts` for an environment.
-    ///
-    /// PlatformIO treats entries without an explicit prefix as POST scripts.
-    /// Values may be provided comma-separated or as a multi-line list.
-    pub fn get_extra_scripts(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        let config = self.get_env_config(env_name)?;
-        match config.get("extra_scripts") {
-            Some(scripts) => Ok(parse_list_values(scripts)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get `board_build.embed_files` for an environment (binary files to embed).
-    pub fn get_embed_files(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        let overrides = self.get_board_overrides(env_name)?;
-        match overrides.get("embed_files") {
-            Some(files) => Ok(parse_lib_deps(files)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get `board_build.embed_txtfiles` for an environment (text files to embed with null terminator).
-    pub fn get_embed_txtfiles(&self, env_name: &str) -> fbuild_core::Result<Vec<String>> {
-        let overrides = self.get_board_overrides(env_name)?;
-        match overrides.get("embed_txtfiles") {
-            Some(files) => Ok(parse_lib_deps(files)),
-            None => Ok(Vec::new()),
-        }
-    }
-
-    /// Get src_dir setting, checking forwarded override and ini config.
-    pub fn get_src_dir(&self, env_name: &str) -> fbuild_core::Result<Option<String>> {
-        // Forwarded PLATFORMIO_SRC_DIR takes precedence.
-        if let Some(env_val) = self.overrides.get_src_dir() {
-            return Ok(Some(env_val.to_string()));
-        }
-
-        let config = self.get_env_config(env_name)?;
-        if let Some(src_dir) = config.get("src_dir") {
-            // Strip inline comments
-            let cleaned = strip_inline_comment(src_dir);
-            if !cleaned.is_empty() {
-                return Ok(Some(cleaned));
-            }
-        }
-
-        // Check [platformio] section
-        if let Some(pio) = self.sections.get("platformio") {
-            if let Some(src_dir) = pio.get("src_dir") {
-                let cleaned = strip_inline_comment(src_dir);
-                if !cleaned.is_empty() {
-                    return Ok(Some(cleaned));
-                }
-            }
-        }
-
-        Ok(None)
-    }
-
-    /// Get board_build.* and board_upload.* overrides from the environment config.
-    pub fn get_board_overrides(
-        &self,
-        env_name: &str,
-    ) -> fbuild_core::Result<HashMap<String, String>> {
-        let config = self.get_env_config(env_name)?;
-        let mut overrides = HashMap::new();
-
-        for (key, value) in config {
-            if let Some(stripped) = key.strip_prefix("board_build.") {
-                overrides.insert(stripped.to_string(), value.clone());
-            } else if let Some(stripped) = key.strip_prefix("board_upload.") {
-                overrides.insert(format!("upload.{}", stripped), value.clone());
-            }
-        }
-
-        Ok(overrides)
-    }
-
-    /// Get the file path this config was loaded from.
-    pub fn path(&self) -> &Path {
-        &self.path
-    }
-
-    // --- Private implementation ---
-
-    /// Parse raw INI content into sections.
-    fn parse_ini(content: &str) -> fbuild_core::Result<HashMap<String, HashMap<String, String>>> {
-        let mut sections: HashMap<String, HashMap<String, String>> = HashMap::new();
-        let mut current_section: Option<String> = None;
-        let mut current_key: Option<String> = None;
-        let mut current_value = String::new();
-
-        for line in content.lines() {
-            let trimmed = line.trim();
-
-            // Skip empty lines and pure comment lines
-            if trimmed.is_empty() || trimmed.starts_with('#') || trimmed.starts_with(';') {
-                // If we're accumulating a multi-line value and hit a blank line,
-                // that ends the multi-line value
-                if trimmed.is_empty() && current_key.is_some() {
-                    if let (Some(ref section), Some(ref key)) = (&current_section, &current_key) {
-                        let section_map = sections.entry(section.clone()).or_default();
-                        section_map.insert(key.clone(), current_value.trim().to_string());
-                        current_key = None;
-                        current_value.clear();
-                    }
-                }
-                continue;
-            }
-
-            // Section header: [section_name]
-            if trimmed.starts_with('[') && trimmed.ends_with(']') {
-                // Flush previous key-value
-                if let (Some(ref section), Some(ref key)) = (&current_section, &current_key) {
-                    let section_map = sections.entry(section.clone()).or_default();
-                    section_map.insert(key.clone(), current_value.trim().to_string());
-                    current_key = None;
-                    current_value.clear();
-                }
-
-                let name = trimmed[1..trimmed.len() - 1].trim().to_string();
-                current_section = Some(name);
-                continue;
-            }
-
-            // Continuation line (starts with whitespace) — append to current value
-            if (line.starts_with(' ') || line.starts_with('\t')) && current_key.is_some() {
-                let val = strip_inline_comment(trimmed);
-                if !current_value.is_empty() {
-                    current_value.push('\n');
-                }
-                current_value.push_str(&val);
-                continue;
-            }
-
-            // Key = value line
-            if let Some(eq_pos) = trimmed.find('=') {
-                // Flush previous key-value
-                if let (Some(ref section), Some(ref key)) = (&current_section, &current_key) {
-                    let section_map = sections.entry(section.clone()).or_default();
-                    section_map.insert(key.clone(), current_value.trim().to_string());
-                }
-
-                let key = trimmed[..eq_pos].trim().to_string();
-                let value = strip_inline_comment(trimmed[eq_pos + 1..].trim());
-
-                current_key = Some(key);
-                current_value = value;
-            }
-        }
-
-        // Flush last key-value
-        if let (Some(ref section), Some(ref key)) = (&current_section, &current_key) {
-            let section_map = sections.entry(section.clone()).or_default();
-            section_map.insert(key.clone(), current_value.trim().to_string());
-        }
-
-        Ok(sections)
-    }
-
-    /// Resolve all `[env:*]` sections with inheritance.
-    fn resolve_all_envs(
-        sections: &HashMap<String, HashMap<String, String>>,
-    ) -> fbuild_core::Result<HashMap<String, HashMap<String, String>>> {
-        let mut resolved = HashMap::new();
-
-        // Find all env:* sections
-        let env_names: Vec<String> = sections
-            .keys()
-            .filter_map(|k| k.strip_prefix("env:").map(|s| s.to_string()))
-            .collect();
-
-        for env_name in &env_names {
-            let config = Self::resolve_env(sections, env_name, &mut HashSet::new())?;
-            // Apply variable substitution
-            let substituted = Self::substitute_vars(sections, &config);
-            resolved.insert(env_name.clone(), substituted);
-        }
-
-        Ok(resolved)
-    }
-
-    /// Resolve a single environment's config, following `extends` chains.
-    fn resolve_env(
-        sections: &HashMap<String, HashMap<String, String>>,
-        env_name: &str,
-        visited: &mut HashSet<String>,
-    ) -> fbuild_core::Result<HashMap<String, String>> {
-        if !visited.insert(env_name.to_string()) {
-            return Err(fbuild_core::FbuildError::ConfigError(format!(
-                "circular extends dependency for env:{}",
-                env_name
-            )));
-        }
-
-        let section_key = format!("env:{}", env_name);
-        let section = sections.get(&section_key).ok_or_else(|| {
-            fbuild_core::FbuildError::ConfigError(format!(
-                "environment '{}' not found in config",
-                env_name
-            ))
-        })?;
-
-        // Start with base [env] section if it exists
-        let mut config: HashMap<String, String> = sections.get("env").cloned().unwrap_or_default();
-
-        // Apply extends (parent values, then child overrides)
-        if let Some(extends) = section.get("extends") {
-            for parent_ref in extends.split(',') {
-                let parent_ref = parent_ref.trim();
-                // extends can reference "env:name" or just "name" (treated as section name)
-                let parent_name = parent_ref.strip_prefix("env:").unwrap_or(parent_ref);
-
-                if sections.contains_key(&format!("env:{}", parent_name)) {
-                    let parent_config = Self::resolve_env(sections, parent_name, visited)?;
-                    // Parent values go first, child overrides later
-                    for (k, v) in parent_config {
-                        config.insert(k, v);
-                    }
-                } else if sections.contains_key(parent_name) {
-                    // Direct section reference (non-env section) — resolve its
-                    // extends chain so inherited keys like lib_deps propagate.
-                    let parent_config = Self::resolve_section(sections, parent_name, visited)?;
-                    for (k, v) in parent_config {
-                        config.insert(k, v);
-                    }
-                }
-            }
-        }
-
-        // Apply this environment's values (override parents)
-        for (k, v) in section {
-            if k != "extends" {
-                config.insert(k.clone(), v.clone());
-            }
-        }
-
-        Ok(config)
-    }
-
-    /// Resolve a non-env section's config, following `extends` chains.
-    fn resolve_section(
-        sections: &HashMap<String, HashMap<String, String>>,
-        section_name: &str,
-        visited: &mut HashSet<String>,
-    ) -> fbuild_core::Result<HashMap<String, String>> {
-        if !visited.insert(format!("section:{}", section_name)) {
-            return Err(fbuild_core::FbuildError::ConfigError(format!(
-                "circular extends dependency for section '{}'",
-                section_name
-            )));
-        }
-
-        let section = sections.get(section_name).cloned().unwrap_or_default();
-        let mut config = HashMap::new();
-
-        // Follow extends chain
-        if let Some(extends) = section.get("extends") {
-            for parent_ref in extends.split(',') {
-                let parent_ref = parent_ref.trim();
-                if sections.contains_key(&format!("env:{}", parent_ref)) {
-                    let parent_config = Self::resolve_env(sections, parent_ref, visited)?;
-                    for (k, v) in parent_config {
-                        config.insert(k, v);
-                    }
-                } else if sections.contains_key(parent_ref) {
-                    let parent_config = Self::resolve_section(sections, parent_ref, visited)?;
-                    for (k, v) in parent_config {
-                        config.insert(k, v);
-                    }
-                }
-            }
-        }
-
-        // Apply this section's own values (override parents)
-        for (k, v) in &section {
-            if k != "extends" {
-                config.insert(k.clone(), v.clone());
-            }
-        }
-
-        Ok(config)
-    }
-
-    /// Apply `${section.key}` variable substitution.
-    fn substitute_vars(
-        sections: &HashMap<String, HashMap<String, String>>,
-        config: &HashMap<String, String>,
-    ) -> HashMap<String, String> {
-        let re = Regex::new(r"\$\{([^}]+)\}").unwrap();
-        let max_depth = 10;
-        let mut result = config.clone();
-
-        for (_key, value) in result.iter_mut() {
-            let mut current = value.clone();
-            for _ in 0..max_depth {
-                let next = re
-                    .replace_all(&current, |caps: &regex::Captures| {
-                        let reference = &caps[1];
-                        resolve_variable(sections, config, reference)
-                    })
-                    .to_string();
-
-                if next == current {
-                    break;
-                }
-                current = next;
-            }
-            *value = current;
-        }
-
-        result
-    }
-}
-
-/// Resolve a variable reference like "section.key" or "env:name.key".
-fn resolve_variable(
-    sections: &HashMap<String, HashMap<String, String>>,
-    current_config: &HashMap<String, String>,
-    reference: &str,
-) -> String {
-    // Try "section.key" format
-    if let Some(dot_pos) = reference.find('.') {
-        let section_ref = &reference[..dot_pos];
-        let key = &reference[dot_pos + 1..];
-
-        // Try env:name format
-        let section_name = if section_ref.starts_with("env:") {
-            section_ref.to_string()
-        } else {
-            // Could be env:name or just a section name
-            let as_env = format!("env:{}", section_ref);
-            if sections.contains_key(&as_env) {
-                as_env
-            } else {
-                section_ref.to_string()
-            }
-        };
-
-        // Look up the key, following the `extends` chain if needed
-        if let Some(val) = resolve_section_key(sections, &section_name, key, &mut HashSet::new()) {
-            return val;
-        }
-    }
-
-    // Try as a key in the current config
-    if let Some(val) = current_config.get(reference) {
-        return val.clone();
-    }
-
-    // Unresolved — return as-is
-    format!("${{{}}}", reference)
-}
-
-/// Look up a key in a section, following `extends` chains.
-fn resolve_section_key(
-    sections: &HashMap<String, HashMap<String, String>>,
-    section_name: &str,
-    key: &str,
-    visited: &mut HashSet<String>,
-) -> Option<String> {
-    if !visited.insert(section_name.to_string()) {
-        return None; // circular extends
-    }
-
-    let section = sections.get(section_name)?;
-
-    // Direct lookup
-    if let Some(val) = section.get(key) {
-        return Some(val.clone());
-    }
-
-    // Follow extends chain
-    if let Some(extends) = section.get("extends") {
-        for parent_ref in extends.split(',') {
-            let parent_ref = parent_ref.trim();
-            // Resolve parent section name (could be "env:name" or bare name)
-            let parent_name = if parent_ref.starts_with("env:") || sections.contains_key(parent_ref)
-            {
-                parent_ref.to_string()
-            } else {
-                let as_env = format!("env:{}", parent_ref);
-                if sections.contains_key(&as_env) {
-                    as_env
-                } else {
-                    parent_ref.to_string()
-                }
-            };
-
-            if let Some(val) = resolve_section_key(sections, &parent_name, key, visited) {
-                return Some(val);
-            }
-        }
-    }
-
-    None
-}
-
-/// Strip inline comments (` ; comment` or ` # comment`).
-/// Be careful not to strip hash/semicolons that are part of values.
-fn strip_inline_comment(s: &str) -> String {
-    // Only strip comments that are preceded by whitespace
-    // This avoids stripping "#include" or URLs with "#"
-    let bytes = s.as_bytes();
-    for i in 1..bytes.len() {
-        if bytes[i - 1] == b' ' && (bytes[i] == b';' || bytes[i] == b'#') {
-            return s[..i].trim().to_string();
-        }
-    }
-    s.trim().to_string()
-}
-
-/// Parse build flags string into a list.
-///
-/// Handles:
-/// - Space-separated flags: `-DFOO -DBAR`
-/// - Multi-line: one flag per line
-/// - `-D FLAG` → `-DFLAG` normalization
-/// - Preserves arguments for `-include`, `-I`, `-L`, etc.
-fn parse_flags(flags_str: &str) -> Vec<String> {
-    let mut result = Vec::new();
-
-    for line in flags_str.lines() {
-        let trimmed = line.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-
-        let chars = trimmed.chars();
-        let mut current = String::new();
-        let mut in_quotes = false;
-        let mut quote_char = ' ';
-
-        for c in chars {
-            match c {
-                '"' | '\'' if !in_quotes => {
-                    in_quotes = true;
-                    quote_char = c;
-                    current.push(c);
-                }
-                c if in_quotes && c == quote_char => {
-                    in_quotes = false;
-                    current.push(c);
-                }
-                ' ' | '\t' if !in_quotes => {
-                    if !current.is_empty() {
-                        result.push(current.clone());
-                        current.clear();
-                    }
-                }
-                _ => {
-                    current.push(c);
-                }
-            }
-        }
-
-        if !current.is_empty() {
-            result.push(current);
-        }
-    }
-
-    // Normalize `-D FLAG` → `-DFLAG`
-    let mut normalized = Vec::new();
-    let mut i = 0;
-    while i < result.len() {
-        if result[i] == "-D" && i + 1 < result.len() {
-            normalized.push(format!("-D{}", result[i + 1]));
-            i += 2;
-        } else {
-            normalized.push(result[i].clone());
-            i += 1;
-        }
-    }
-
-    normalized
-}
-
-/// Parse library dependencies from a multi-line or comma-separated string.
-fn parse_lib_deps(deps_str: &str) -> Vec<String> {
-    let mut result = Vec::new();
-
-    for line in deps_str.lines() {
-        for dep in line.split(',') {
-            let trimmed = dep.trim();
-            if !trimmed.is_empty() {
-                result.push(trimmed.to_string());
-            }
-        }
-    }
-
-    result
-}
-
-/// Parse a generic multi-value option from a multi-line or comma-separated string.
-fn parse_list_values(value: &str) -> Vec<String> {
-    let mut result = Vec::new();
-
-    for line in value.lines() {
-        for item in line.split(',') {
-            let trimmed = item.trim();
-            if !trimmed.is_empty() {
-                result.push(trimmed.to_string());
-            }
-        }
-    }
-
-    result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::BTreeMap;
-    use std::io::Write;
-    use tempfile::NamedTempFile;
-
-    fn write_ini(content: &str) -> NamedTempFile {
-        let mut f = NamedTempFile::new().unwrap();
-        f.write_all(content.as_bytes()).unwrap();
-        f.flush().unwrap();
-        f
-    }
-
-    fn overrides(pairs: &[(&str, &str)]) -> crate::pio_env::PioEnvOverrides {
-        crate::pio_env::PioEnvOverrides::from_map(
-            pairs
-                .iter()
-                .map(|(k, v)| ((*k).to_string(), (*v).to_string()))
-                .collect::<BTreeMap<_, _>>(),
-        )
-    }
-
-    #[test]
-    fn test_init_with_valid_file() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(config.get_environments(), vec!["uno"]);
-    }
-
-    #[test]
-    fn test_init_with_nonexistent_file() {
-        let result = PlatformIOConfig::from_path(Path::new("/nonexistent/platformio.ini"));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_get_environments_multiple() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-
-[env:esp32]
-platform = espressif32
-board = esp32dev
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let envs = config.get_environments();
-        assert_eq!(envs.len(), 2);
-        assert!(envs.contains(&"uno"));
-        assert!(envs.contains(&"esp32"));
-    }
-
-    #[test]
-    fn test_get_environments_empty() {
-        let f = write_ini("[platformio]\ndefault_envs = \n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert!(config.get_environments().is_empty());
-    }
-
-    #[test]
-    fn test_get_env_config_valid() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let env = config.get_env_config("uno").unwrap();
-        assert_eq!(env.get("platform").unwrap(), "atmelavr");
-        assert_eq!(env.get("board").unwrap(), "uno");
-    }
-
-    #[test]
-    fn test_get_env_config_nonexistent() {
-        let f = write_ini("[env:uno]\nplatform = atmelavr\nboard = uno\nframework = arduino\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert!(config.get_env_config("nonexistent").is_err());
-    }
-
-    #[test]
-    fn test_get_env_config_with_base_env_inheritance() {
-        let f = write_ini(
-            "\
-[env]
-framework = arduino
-
-[env:uno]
-platform = atmelavr
-board = uno
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let env = config.get_env_config("uno").unwrap();
-        assert_eq!(env.get("framework").unwrap(), "arduino");
-        assert_eq!(env.get("platform").unwrap(), "atmelavr");
-    }
-
-    #[test]
-    fn test_get_build_flags_present() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-build_flags = -DFOO -DBAR
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let flags = config.get_build_flags("uno").unwrap();
-        assert_eq!(flags, vec!["-DFOO", "-DBAR"]);
-    }
-
-    #[test]
-    fn test_get_build_flags_absent() {
-        let f = write_ini("[env:uno]\nplatform = atmelavr\nboard = uno\nframework = arduino\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let flags = config.get_build_flags("uno").unwrap();
-        assert!(flags.is_empty());
-    }
-
-    #[test]
-    fn test_get_build_flags_prefers_forwarded_override() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-build_flags = -DINI=1
-",
-        );
-        let config = PlatformIOConfig::from_path_with_overrides(
-            f.path(),
-            overrides(&[("PLATFORMIO_BUILD_FLAGS", "-DOVERRIDE=1 -DFAST=1")]),
-        )
-        .unwrap();
-        assert_eq!(
-            config.get_build_flags("uno").unwrap(),
-            vec!["-DOVERRIDE=1", "-DFAST=1"]
-        );
-    }
-
-    #[test]
-    fn test_get_build_flags_multiline() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-build_flags =
-    -DFOO
-    -DBAR
-    -DBAZ
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let flags = config.get_build_flags("uno").unwrap();
-        assert_eq!(flags, vec!["-DFOO", "-DBAR", "-DBAZ"]);
-    }
-
-    #[test]
-    fn test_get_build_flags_d_space_normalization() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-build_flags = -D FOO -D BAR
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let flags = config.get_build_flags("uno").unwrap();
-        assert_eq!(flags, vec!["-DFOO", "-DBAR"]);
-    }
-
-    #[test]
-    fn test_get_build_src_flags_prefers_forwarded_override() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-build_src_flags = -DINI_SRC
-",
-        );
-        let config = PlatformIOConfig::from_path_with_overrides(
-            f.path(),
-            overrides(&[("PLATFORMIO_BUILD_SRC_FLAGS", "-DOVERRIDE_SRC=1")]),
-        )
-        .unwrap();
-        assert_eq!(
-            config.get_build_src_flags("uno").unwrap(),
-            vec!["-DOVERRIDE_SRC=1"]
-        );
-    }
-
-    #[test]
-    fn test_get_extra_scripts_multiline_and_comma_separated() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-extra_scripts =
-    pre:scripts/pre.py, scripts/post.py
-    post:scripts/after.py
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let scripts = config.get_extra_scripts("uno").unwrap();
-        assert_eq!(
-            scripts,
-            vec![
-                "pre:scripts/pre.py",
-                "scripts/post.py",
-                "post:scripts/after.py"
-            ]
-        );
-    }
-
-    #[test]
-    fn test_get_lib_deps_present() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-lib_deps =
-    FastLED
-    ArduinoJson
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let deps = config.get_lib_deps("uno").unwrap();
-        assert_eq!(deps, vec!["FastLED", "ArduinoJson"]);
-    }
-
-    #[test]
-    fn test_get_lib_deps_comma_separated() {
-        let f = write_ini(
-            "\
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-lib_deps = FastLED, ArduinoJson
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let deps = config.get_lib_deps("uno").unwrap();
-        assert_eq!(deps, vec!["FastLED", "ArduinoJson"]);
-    }
-
-    #[test]
-    fn test_get_lib_deps_absent() {
-        let f = write_ini("[env:uno]\nplatform = atmelavr\nboard = uno\nframework = arduino\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let deps = config.get_lib_deps("uno").unwrap();
-        assert!(deps.is_empty());
-    }
-
-    #[test]
-    fn test_has_environment() {
-        let f = write_ini("[env:uno]\nplatform = atmelavr\nboard = uno\nframework = arduino\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert!(config.has_environment("uno"));
-        assert!(!config.has_environment("esp32"));
-    }
-
-    #[test]
-    fn test_get_default_environment_explicit() {
-        let f = write_ini(
-            "\
-[platformio]
-default_envs = esp32
-
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-
-[env:esp32]
-platform = espressif32
-board = esp32dev
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(config.get_default_environment(), Some("esp32"));
-    }
-
-    #[test]
-    fn test_get_default_environment_prefers_forwarded_override() {
-        let f = write_ini(
-            "\
-[platformio]
-default_envs = esp32
-
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-
-[env:esp32]
-platform = espressif32
-board = esp32dev
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path_with_overrides(
-            f.path(),
-            overrides(&[("PLATFORMIO_DEFAULT_ENVS", "uno")]),
-        )
-        .unwrap();
-        assert_eq!(config.get_default_environment(), Some("uno"));
-    }
-
-    #[test]
-    fn test_get_default_environment_first_fallback() {
-        let f = write_ini(
-            "\
-[env:alpha]
-platform = atmelavr
-board = uno
-framework = arduino
-
-[env:beta]
-platform = espressif32
-board = esp32dev
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        // Should return first alphabetically
-        assert_eq!(config.get_default_environment(), Some("alpha"));
-    }
-
-    #[test]
-    fn test_get_default_environment_none() {
-        let f = write_ini("[platformio]\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(config.get_default_environment(), None);
-    }
-
-    #[test]
-    fn test_extends_inheritance() {
-        let f = write_ini(
-            "\
-[env:base]
-platform = atmelavr
-framework = arduino
-build_flags = -DBASE
-
-[env:child]
-extends = env:base
-board = uno
-build_flags = -DCHILD
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let env = config.get_env_config("child").unwrap();
-        assert_eq!(env.get("platform").unwrap(), "atmelavr");
-        assert_eq!(env.get("framework").unwrap(), "arduino");
-        assert_eq!(env.get("board").unwrap(), "uno");
-        // Child overrides parent's build_flags
-        assert_eq!(env.get("build_flags").unwrap(), "-DCHILD");
-    }
-
-    #[test]
-    fn test_get_src_dir_from_ini() {
-        let f = write_ini(
-            "\
-[platformio]
-src_dir = custom_src
-
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(
-            config.get_src_dir("uno").unwrap(),
-            Some("custom_src".to_string())
-        );
-    }
-
-    #[test]
-    fn test_get_src_dir_returns_none_when_not_set() {
-        let f = write_ini("[env:uno]\nplatform = atmelavr\nboard = uno\nframework = arduino\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(config.get_src_dir("uno").unwrap(), None);
-    }
-
-    #[test]
-    fn test_get_src_dir_with_inline_comment() {
-        let f = write_ini(
-            "\
-[platformio]
-src_dir = custom_src ; this is the source dir
-
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(
-            config.get_src_dir("uno").unwrap(),
-            Some("custom_src".to_string())
-        );
-    }
-
-    #[test]
-    fn test_get_src_dir_prefers_forwarded_override() {
-        let f = write_ini(
-            "\
-[platformio]
-src_dir = ini_src
-
-[env:uno]
-platform = atmelavr
-board = uno
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path_with_overrides(
-            f.path(),
-            overrides(&[("PLATFORMIO_SRC_DIR", "override_src")]),
-        )
-        .unwrap();
-        assert_eq!(
-            config.get_src_dir("uno").unwrap(),
-            Some("override_src".to_string())
-        );
-    }
-
-    #[test]
-    fn test_real_world_config() {
-        let f = write_ini(
-            "\
-[platformio]
-default_envs = esp32dev
-
-[env]
-framework = arduino
-
-[env:esp32dev]
-platform = espressif32
-board = esp32dev
-build_flags =
-    -DFASTLED_ESP32
-    -DCORE_DEBUG_LEVEL=0
-lib_deps =
-    FastLED
-    ArduinoJson
-
-[env:uno]
-platform = atmelavr
-board = uno
-build_flags = -DFASTLED_AVR
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-
-        // Default env
-        assert_eq!(config.get_default_environment(), Some("esp32dev"));
-
-        // ESP32 config
-        let esp = config.get_env_config("esp32dev").unwrap();
-        assert_eq!(esp.get("platform").unwrap(), "espressif32");
-        assert_eq!(esp.get("framework").unwrap(), "arduino"); // inherited from [env]
-
-        let esp_flags = config.get_build_flags("esp32dev").unwrap();
-        assert_eq!(esp_flags, vec!["-DFASTLED_ESP32", "-DCORE_DEBUG_LEVEL=0"]);
-
-        let esp_deps = config.get_lib_deps("esp32dev").unwrap();
-        assert_eq!(esp_deps, vec!["FastLED", "ArduinoJson"]);
-
-        // Uno config
-        let uno = config.get_env_config("uno").unwrap();
-        assert_eq!(uno.get("platform").unwrap(), "atmelavr");
-        assert_eq!(uno.get("framework").unwrap(), "arduino"); // inherited from [env]
-
-        let uno_flags = config.get_build_flags("uno").unwrap();
-        assert_eq!(uno_flags, vec!["-DFASTLED_AVR"]);
-    }
-
-    #[test]
-    fn test_strip_inline_comment() {
-        assert_eq!(strip_inline_comment("value ; comment"), "value");
-        assert_eq!(strip_inline_comment("value # comment"), "value");
-        assert_eq!(strip_inline_comment("value"), "value");
-        assert_eq!(strip_inline_comment("#include <foo>"), "#include <foo>");
-    }
-
-    #[test]
-    fn test_parse_flags() {
-        assert_eq!(parse_flags("-DFOO -DBAR"), vec!["-DFOO", "-DBAR"]);
-        assert_eq!(parse_flags("-D FOO -D BAR"), vec!["-DFOO", "-DBAR"]);
-        assert_eq!(
-            parse_flags("-DFOO\n-DBAR\n-DBAZ"),
-            vec!["-DFOO", "-DBAR", "-DBAZ"]
-        );
-    }
-
-    #[test]
-    fn test_parse_lib_deps() {
-        assert_eq!(
-            parse_lib_deps("FastLED, ArduinoJson"),
-            vec!["FastLED", "ArduinoJson"]
-        );
-        assert_eq!(
-            parse_lib_deps("FastLED\nArduinoJson"),
-            vec!["FastLED", "ArduinoJson"]
-        );
-    }
-
-    #[test]
-    fn test_get_lib_ignore_present() {
-        let f = write_ini(
-            "\
-[env:esp32dev]
-platform = espressif32
-board = esp32dev
-framework = arduino
-lib_ignore =
-    WiFi
-    FS
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let ignore = config.get_lib_ignore("esp32dev").unwrap();
-        assert_eq!(ignore, vec!["WiFi", "FS"]);
-    }
-
-    #[test]
-    fn test_get_lib_ignore_absent() {
-        let f = write_ini("[env:uno]\nplatform = atmelavr\nboard = uno\nframework = arduino\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let ignore = config.get_lib_ignore("uno").unwrap();
-        assert!(ignore.is_empty());
-    }
-
-    #[test]
-    fn test_variable_substitution_follows_extends_chain() {
-        // Mirrors NightDriverStrip pattern: [base] defines build_flags,
-        // [dev_esp32] extends base (no build_flags), [env:demo] references
-        // ${dev_esp32.build_flags} — must resolve through the extends chain.
-        let f = write_ini(
-            "\
-[base]
-build_flags = -std=gnu++2a
-              -Ofast
-
-[dev_esp32]
-extends = base
-board = esp32dev
-
-[env:demo]
-extends = dev_esp32
-platform = espressif32
-framework = arduino
-build_flags = -DDEMO=1
-              ${dev_esp32.build_flags}
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let flags = config.get_build_flags("demo").unwrap();
-        assert!(
-            flags.contains(&"-DDEMO=1".to_string()),
-            "should contain -DDEMO=1, got: {:?}",
-            flags
-        );
-        assert!(
-            flags.contains(&"-std=gnu++2a".to_string()),
-            "should resolve ${{dev_esp32.build_flags}} through extends chain to [base], got: {:?}",
-            flags
-        );
-        assert!(
-            flags.contains(&"-Ofast".to_string()),
-            "should contain -Ofast from [base], got: {:?}",
-            flags
-        );
-        // Must NOT contain the literal unresolved variable
-        let raw = config.get_env_config("demo").unwrap();
-        let raw_flags = raw.get("build_flags").unwrap();
-        assert!(
-            !raw_flags.contains("${"),
-            "build_flags should not contain unresolved variables, got: {}",
-            raw_flags
-        );
-    }
-
-    #[test]
-    fn test_non_env_extends_inherits_lib_deps() {
-        // [env:demo] extends [dev_esp32] extends [base].
-        // lib_deps is only on [base] — must propagate through.
-        let f = write_ini(
-            "\
-[base]
-lib_deps = fastled/FastLED@^3.7.8
-           bblanchon/ArduinoJson@^7.0.0
-
-[dev_esp32]
-extends = base
-board = esp32dev
-
-[env:demo]
-extends = dev_esp32
-platform = espressif32
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let deps = config.get_lib_deps("demo").unwrap();
-        assert!(
-            deps.iter().any(|d| d.contains("FastLED")),
-            "should inherit lib_deps from [base] via [dev_esp32], got: {:?}",
-            deps
-        );
-        assert!(
-            deps.iter().any(|d| d.contains("ArduinoJson")),
-            "should inherit ArduinoJson from [base], got: {:?}",
-            deps
-        );
-    }
-
-    #[test]
-    fn test_get_embed_files() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = espressif32
-board = esp32dev
-framework = arduino
-board_build.embed_files = site/dist/index.html.gz
-                          site/dist/favicon.ico.gz
-board_build.embed_txtfiles = config/timezones.json
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        let embed = config.get_embed_files("demo").unwrap();
-        assert_eq!(embed.len(), 2);
-        assert!(embed.contains(&"site/dist/index.html.gz".to_string()));
-        assert!(embed.contains(&"site/dist/favicon.ico.gz".to_string()));
-
-        let txt = config.get_embed_txtfiles("demo").unwrap();
-        assert_eq!(txt.len(), 1);
-        assert_eq!(txt[0], "config/timezones.json");
-    }
-
-    #[test]
-    fn test_get_embed_files_empty() {
-        let f = write_ini("[env:uno]\nplatform = atmelavr\nboard = uno\nframework = arduino\n");
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert!(config.get_embed_files("uno").unwrap().is_empty());
-        assert!(config.get_embed_txtfiles("uno").unwrap().is_empty());
-    }
-
-    #[test]
-    fn test_get_source_filter_prefers_build_src_filter() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = espressif32
-board = esp32dev
-framework = arduino
-src_filter = +<legacy.cpp>
-build_src_filter =
-    +<*>
-    -<generated/>
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(
-            config.get_source_filter("demo").unwrap(),
-            Some("+<*>\n-<generated/>".to_string())
-        );
-    }
-
-    #[test]
-    fn test_get_source_filter_falls_back_to_src_filter() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = ststm32
-board = bluepill_f103c8
-framework = stm32cube
-src_filter =
-    +<main.cpp>
-    -<legacy/>
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(
-            config.get_source_filter("demo").unwrap(),
-            Some("+<main.cpp>\n-<legacy/>".to_string())
-        );
-    }
-
-    #[test]
-    fn test_get_build_unflags_prefers_env_override() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = espressif32
-board = esp32dev
-framework = arduino
-build_unflags = -std=gnu++11
-",
-        );
-        let overrides = crate::pio_env::PioEnvOverrides::from_map(
-            [(
-                "PLATFORMIO_BUILD_UNFLAGS".to_string(),
-                "-std=gnu++17 -DDEBUG".to_string(),
-            )]
-            .into_iter()
-            .collect(),
-        );
-        let config = PlatformIOConfig::from_path_with_overrides(f.path(), overrides).unwrap();
-        assert_eq!(
-            config.get_build_unflags("demo").unwrap(),
-            vec!["-std=gnu++17", "-DDEBUG"]
-        );
-    }
-
-    #[test]
-    fn test_get_build_unflags_from_ini() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = ststm32
-board = bluepill_f103c8
-framework = stm32cube
-build_unflags =
-    -Os
-    -DDEBUG
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(
-            config.get_build_unflags("demo").unwrap(),
-            vec!["-Os", "-DDEBUG"]
-        );
-    }
-
-    #[test]
-    fn test_get_build_type_defaults_to_release() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = atmelavr
-board = uno
-framework = arduino
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(config.get_build_type("demo").unwrap(), "release");
-    }
-
-    #[test]
-    fn test_get_build_type_reads_debug() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = espressif32
-board = esp32dev
-framework = arduino
-build_type = debug
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(config.get_build_type("demo").unwrap(), "debug");
-    }
-
-    #[test]
-    fn test_get_debug_build_flags_uses_platformio_defaults() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = espressif32
-board = esp32dev
-framework = arduino
-build_type = debug
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(
-            config.get_debug_build_flags("demo").unwrap(),
-            vec!["-Og", "-g2", "-ggdb2"]
-        );
-    }
-
-    #[test]
-    fn test_get_debug_build_flags_reads_ini_override() {
-        let f = write_ini(
-            "\
-[env:demo]
-platform = espressif32
-board = esp32dev
-framework = arduino
-build_type = debug
-debug_build_flags = -Og -g3
-",
-        );
-        let config = PlatformIOConfig::from_path(f.path()).unwrap();
-        assert_eq!(
-            config.get_debug_build_flags("demo").unwrap(),
-            vec!["-Og", "-g3"]
-        );
-    }
-}
